Individual Project [21_22]/FYP-main/pyproject.toml
[tool.poetry]
name = "fyp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [
    { include = "fyp", from = "src" },
]

[tool.poetry.scripts]
CollectNeutralSeedUsers = 'fyp.main:collect_neutral_seed_users'
CollectOtherSeedUsers = 'fyp.main:collect_other_seed_users'
GenerateDatabaseKey = 'fyp.FernetCrypto:generate_key'
SnowballSample = 'fyp.main:snowball_sample_db'
EncryptFile = 'fyp.main:encrypt_file'

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
requests = "^2.26.0"
python-dotenv = "^0.19.2"
numpy = "^1.21.5"
cryptography = "^36.0.1"
peewee = "^3.14.8"
ptyprocess = "^0.7.0"
networkx = "^2.6.3"
tqdm = "^4.63.0"
matplotlib = "^3.5.1"
scipy = "^1.8.0"
ijson = "^3.1.4"
msgpack = "^1.0.3"
nltk = "^3.7"
seaborn = "^0.11.2"
tabulate = "^0.8.9"
sklearn = "^0.0"
scikit-learn = "^1.0.2"
contractions = "^0.1.72"
timer = "^0.2.2"
mlxtend = "^0.19.0"
emoji = "^1.7.0"
word2number = "^1.1"
cityhash = "^0.4.1"
statsmodels = "^0.13.2"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^21.12b0"
isort = "^5.10.1"
pre-commit = "^2.16.0"
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
